refactor: componentize payment reducer
