+++ renderer/reducers/payment/constants.js
@@ -1,0 +1,14 @@
+export const GET_PAYMENTS = 'GET_PAYMENTS'
+export const RECEIVE_PAYMENTS = 'RECEIVE_PAYMENTS'
+export const SEND_PAYMENT = 'SEND_PAYMENT'
+export const PAYMENT_SUCCESSFUL = 'PAYMENT_SUCCESSFUL'
+export const PAYMENT_FAILED = 'PAYMENT_FAILED'
+export const DECREASE_PAYMENT_RETRIES = 'DECREASE_PAYMENT_RETRIES'
+
+export const PAYMENT_STATUS_SENDING = 'sending'
+export const PAYMENT_STATUS_SUCCESSFUL = 'successful'
+export const PAYMENT_STATUS_FAILED = 'failed'
+
+export const DEFAULT_CLTV_DELTA = 43
+export const KEYSEND_PREIMAGE_TYPE = '5482373484'
+export const PREIMAGE_BYTE_LENGTH = 32
