+++ renderer/reducers/payment/index.js
@@ -1,0 +1,6 @@
+import paymentReducer from './reducer'
+
+export default paymentReducer
+export paymentSelectors from './selectors'
+export * from './constants'
+export * from './reducer'
