+++ renderer/reducers/payment/messages.js
@@ -1,0 +1,8 @@
+import { defineMessages } from 'react-intl'
+
+// use pattern [reducer]_message
+
+/* eslint-disable max-len */
+export default defineMessages({
+  payment_send_error: 'Unable to send payment: Invalid invoice (no payment hash)',
+})
