+++ renderer/reducers/payment/reducer.js
@@ -1,28 +1,44 @@
 import config from 'config'
 import { randomBytes, createHash } from 'crypto'
-import { createSelector } from 'reselect'
 import uniqBy from 'lodash/uniqBy'
 import find from 'lodash/find'
 import createReducer from '@zap/utils/createReducer'
 import errorToUserFriendly from '@zap/utils/userFriendlyErrors'
 import { getIntl } from '@zap/i18n'
-import { decodePayReq, getNodeAlias, getTag, isPubkey } from '@zap/utils/crypto'
-import { convert } from '@zap/utils/btc'
+import { decodePayReq, getTag, isPubkey } from '@zap/utils/crypto'
 import delay from '@zap/utils/delay'
 import genId from '@zap/utils/genId'
 import { mainLog } from '@zap/utils/log'
 import { CoinBig } from '@zap/utils/coin'
 import { grpc } from 'workers'
-import { fetchBalance } from './balance'
-import { fetchChannels } from './channels'
-import { infoSelectors } from './info'
-import { networkSelectors } from './network'
-import { showError } from './notification'
+import { fetchBalance } from 'reducers/balance'
+import { fetchChannels } from 'reducers/channels'
+import { infoSelectors } from 'reducers/info'
+import { showError } from 'reducers/notification'
 import messages from './messages'
+import paymentSelectors from './selectors'
+import * as constants from './constants'
 
-export const DEFAULT_CLTV_DELTA = 43
-export const KEYSEND_PREIMAGE_TYPE = '5482373484'
-export const PREIMAGE_BYTE_LENGTH = 32
+const {
+  GET_PAYMENTS,
+  RECEIVE_PAYMENTS,
+  SEND_PAYMENT,
+  PAYMENT_SUCCESSFUL,
+  PAYMENT_FAILED,
+  DECREASE_PAYMENT_RETRIES,
+  PAYMENT_STATUS_SENDING,
+  PAYMENT_STATUS_SUCCESSFUL,
+  PAYMENT_STATUS_FAILED,
+  DEFAULT_CLTV_DELTA,
+  KEYSEND_PREIMAGE_TYPE,
+  PREIMAGE_BYTE_LENGTH,
+} = constants
+
+// ------------------------------------
+// Constants
+// ------------------------------------
+
+const PAYMENT_TIMEOUT = config.invoices.paymentTimeout
 
 // ------------------------------------
 // Initial State
@@ -32,86 +48,6 @@ const initialState = {
   paymentLoading: false,
   payments: [],
   paymentsSending: [],
-}
-
-// ------------------------------------
-// Constants
-// ------------------------------------
-
-export const SET_PAYMENT = 'SET_PAYMENT'
-export const GET_PAYMENTS = 'GET_PAYMENTS'
-export const RECEIVE_PAYMENTS = 'RECEIVE_PAYMENTS'
-export const SEND_PAYMENT = 'SEND_PAYMENT'
-export const PAYMENT_SUCCESSFUL = 'PAYMENT_SUCCESSFUL'
-export const PAYMENT_FAILED = 'PAYMENT_FAILED'
-export const DECREASE_PAYMENT_RETRIES = 'DECREASE_PAYMENT_RETRIES'
-
-const PAYMENT_STATUS_SENDING = 'sending'
-const PAYMENT_STATUS_SUCCESSFUL = 'successful'
-const PAYMENT_STATUS_FAILED = 'failed'
-
-const PAYMENT_TIMEOUT = config.invoices.paymentTimeout
-
-// ------------------------------------
-// Helpers
-// ------------------------------------
-
-/**
- * decoratePayment - Decorate payment object with custom/computed properties.
- *
- * @param  {object} payment Payment
- * @param  {Array} nodes Nodes
- * @returns {object} Decorated payment
- */
-const decoratePayment = (payment, nodes = []) => {
-  // Add basic type information.
-  const decoration = {
-    type: 'payment',
-  }
-
-  // Older versions of lnd provided the sat amount in `value`.
-  // This is now deprecated in favor of `valueSat` and `valueMsat`.
-  // Patch data returned from older clients to match the current format for consistency.
-  const { value } = payment
-  if (value && (!payment.valueSat || !payment.valueMsat)) {
-    Object.assign(decoration, {
-      valueSat: value,
-      valueMsat: convert('sats', 'msats', value),
-    })
-  }
-
-  // Convert the preimage to a hex string.
-  if (!payment.paymentPreimage) {
-    decoration.paymentPreimage =
-      payment.rPreimage && Buffer.from(payment.rPreimage, 'hex').toString('hex')
-  }
-
-  // Convert the preimage to a hex string.
-  if (!payment.paymentHash) {
-    decoration.paymentHash = payment.rHash && Buffer.from(payment.rHash, 'hex').toString('hex')
-  }
-
-  // First try to get the pubkey from payment.path
-  let pubkey = payment.path && payment.path[payment.path.length - 1]
-
-  // If we don't have a pubkey, try to get it from the payment request.
-  if (!pubkey && payment.paymentRequest) {
-    const paymentRequest = decodePayReq(payment.paymentRequest)
-    pubkey = paymentRequest.payeeNodeKey
-  }
-
-  // Try to add some info about the destination of the payment.
-  if (pubkey) {
-    Object.assign(decoration, {
-      destNodePubkey: pubkey,
-      destNodeAlias: getNodeAlias(pubkey, nodes),
-    })
-  }
-
-  return {
-    ...payment,
-    ...decoration,
-  }
 }
 
 // ------------------------------------
@@ -128,6 +64,29 @@ export function getPayments() {
     type: GET_PAYMENTS,
   }
 }
+/**
+ * receivePayments - Fetch payments success callback.
+ *
+ * @param {Array} payments list of payments.
+ * @returns {Function} Thunk
+ */
+export const receivePayments = payments => dispatch => {
+  dispatch({
+    type: RECEIVE_PAYMENTS,
+    payments,
+  })
+}
+
+/**
+ * decPaymentRetry - Decrement payment request retry count.
+ *
+ * @param {string} paymentId Internal id of payment whose retry count to decrease
+ * @returns {object} Action
+ */
+const decPaymentRetry = paymentId => ({
+  type: DECREASE_PAYMENT_RETRIES,
+  paymentId,
+})
 
 /**
  * sendPayment - After initiating a lightning payment, store details of it in paymentSending array.
@@ -166,25 +125,96 @@ export const fetchPayments = () => async dispatch => {
 }
 
 /**
- * receivePayments - Fetch payments success callback.
+ * updatePayment - Updates specified payment request.
  *
- * @param {Array} payments list of payments.
+ * @param {string} paymentRequest Payment request
  * @returns {Function} Thunk
  */
-export const receivePayments = payments => dispatch => {
-  dispatch({ type: RECEIVE_PAYMENTS, payments })
+export const updatePayment = paymentRequest => async dispatch => {
+  const { payments } = await grpc.services.Lightning.listPayments()
+  const payment = payments.find(p => p.paymentRequest === paymentRequest)
+  if (payment) {
+    dispatch(receivePayments([payment]))
+  }
 }
 
 /**
- * decPaymentRetry - Decrement payment request retry count.
+ * paymentSuccessful - Success handler for payInvoice.
  *
- * @param {string} paymentId Internal id of payment whose retry count to decrease
- * @returns {object} Action
+ * @param {{paymentRequest}} paymentRequest Payment request
+ * @returns {Function} Thunk
  */
-const decPaymentRetry = paymentId => ({
-  type: DECREASE_PAYMENT_RETRIES,
-  paymentId,
-})
+export const paymentSuccessful = ({ paymentId }) => async (dispatch, getState) => {
+  const paymentSending = find(paymentSelectors.paymentsSendingSelector(getState()), { paymentId })
+
+  // If we found a related entry in paymentsSending, gracefully remove it and handle as success case.
+  if (paymentSending) {
+    const { creationDate, paymentRequest } = paymentSending
+
+    // Ensure payment stays in sending state for at least 2 seconds.
+    await delay(2000 - (Date.now() - creationDate * 1000))
+
+    // Mark the payment as successful.
+    dispatch({ type: PAYMENT_SUCCESSFUL, paymentId })
+
+    // Wait for another second.
+    await delay(1500)
+
+    dispatch(updatePayment(paymentRequest))
+  }
+
+  // Fetch new balance.
+  dispatch(fetchBalance())
+
+  // Fetch updated channels.
+  dispatch(fetchChannels())
+}
+
+/**
+ * paymentFailed - Error handler for payInvoice.
+ *
+ * @param {Error} error Error
+ * @param {object} details Failed payment details
+ *
+ * @returns {Function} Thunk
+ */
+export const paymentFailed = (error, { paymentId }) => async (dispatch, getState) => {
+  const paymentSending = find(paymentSelectors.paymentsSendingSelector(getState()), { paymentId })
+
+  // errors that trigger retry mechanism
+  const RETRIABLE_ERRORS = [
+    'payment attempt not completed before timeout', // ErrPaymentAttemptTimeout
+    'unable to find a path to destination', // ErrNoPathFound
+    'target not found', // ErrTargetNotInNetwork
+    'FAILED_NO_ROUTE',
+    'FAILED_ERROR',
+    'FAILED_TIMEOUT',
+    'TERMINATED_EARLY', // Triggered if sendPayment aborts without giveing a proper response.
+  ]
+
+  // If we found a related entery in paymentsSending, gracefully remove it and handle as error case.
+  if (paymentSending) {
+    const { creationDate, paymentRequest, remainingRetries, maxRetries } = paymentSending
+    // if we have retries left and error is eligible for retry - rebroadcast payment
+    if (remainingRetries && RETRIABLE_ERRORS.includes(error)) {
+      const data = {
+        ...paymentSending,
+        payReq: paymentRequest,
+        originalPaymentId: paymentId,
+      }
+      const retryIndex = maxRetries - remainingRetries + 1
+      // add increasing delay
+      await delay(config.invoices.baseRetryDelay * retryIndex * retryIndex)
+      dispatch(payInvoice(data))
+    } else {
+      // Ensure payment stays in sending state for at least 2 seconds.
+      await delay(2000 - (Date.now() - creationDate * 1000))
+
+      // Mark the payment as failed.
+      dispatch({ type: PAYMENT_FAILED, paymentId, error: errorToUserFriendly(error) })
+    }
+  }
+}
 
 /**
  * payInvoice - Pay a lightniung invoice.
@@ -332,98 +362,6 @@ export const payInvoice = ({
   }
 }
 
-/**
- * updatePayment - Updates specified payment request.
- *
- * @param {string} paymentRequest Payment request
- * @returns {Function} Thunk
- */
-export const updatePayment = paymentRequest => async dispatch => {
-  const { payments } = await grpc.services.Lightning.listPayments()
-  const payment = payments.find(p => p.paymentRequest === paymentRequest)
-  if (payment) {
-    dispatch(receivePayments([payment]))
-  }
-}
-
-/**
- * paymentSuccessful - Success handler for payInvoice.
- *
- * @param {{paymentRequest}} paymentRequest Payment request
- * @returns {Function} Thunk
- */
-export const paymentSuccessful = ({ paymentId }) => async (dispatch, getState) => {
-  const paymentSending = find(paymentsSendingSelector(getState()), { paymentId })
-
-  // If we found a related entry in paymentsSending, gracefully remove it and handle as success case.
-  if (paymentSending) {
-    const { creationDate, paymentRequest } = paymentSending
-
-    // Ensure payment stays in sending state for at least 2 seconds.
-    await delay(2000 - (Date.now() - creationDate * 1000))
-
-    // Mark the payment as successful.
-    dispatch({ type: PAYMENT_SUCCESSFUL, paymentId })
-
-    // Wait for another second.
-    await delay(1500)
-
-    dispatch(updatePayment(paymentRequest))
-  }
-
-  // Fetch new balance.
-  dispatch(fetchBalance())
-
-  // Fetch updated channels.
-  dispatch(fetchChannels())
-}
-
-/**
- * paymentFailed - Error handler for payInvoice.
- *
- * @param {Error} error Error
- * @param {object} details Failed payment details
- *
- * @returns {Function} Thunk
- */
-export const paymentFailed = (error, { paymentId }) => async (dispatch, getState) => {
-  const paymentSending = find(paymentsSendingSelector(getState()), { paymentId })
-
-  // errors that trigger retry mechanism
-  const RETRIABLE_ERRORS = [
-    'payment attempt not completed before timeout', // ErrPaymentAttemptTimeout
-    'unable to find a path to destination', // ErrNoPathFound
-    'target not found', // ErrTargetNotInNetwork
-    'FAILED_NO_ROUTE',
-    'FAILED_ERROR',
-    'FAILED_TIMEOUT',
-    'TERMINATED_EARLY', // Triggered if sendPayment aborts without giveing a proper response.
-  ]
-
-  // If we found a related entery in paymentsSending, gracefully remove it and handle as error case.
-  if (paymentSending) {
-    const { creationDate, paymentRequest, remainingRetries, maxRetries } = paymentSending
-    // if we have retries left and error is eligible for retry - rebroadcast payment
-    if (remainingRetries && RETRIABLE_ERRORS.includes(error)) {
-      const data = {
-        ...paymentSending,
-        payReq: paymentRequest,
-        originalPaymentId: paymentId,
-      }
-      const retryIndex = maxRetries - remainingRetries + 1
-      // add increasing delay
-      await delay(config.invoices.baseRetryDelay * retryIndex * retryIndex)
-      dispatch(payInvoice(data))
-    } else {
-      // Ensure payment stays in sending state for at least 2 seconds.
-      await delay(2000 - (Date.now() - creationDate * 1000))
-
-      // Mark the payment as failed.
-      dispatch({ type: PAYMENT_FAILED, paymentId, error: errorToUserFriendly(error) })
-    }
-  }
-}
-
 // ------------------------------------
 // Action Handlers
 // ------------------------------------
@@ -442,7 +380,6 @@ const ACTION_HANDLERS = {
   [SEND_PAYMENT]: (state, { payment }) => {
     state.paymentsSending.push(payment)
   },
-
   [DECREASE_PAYMENT_RETRIES]: (state, { paymentId }) => {
     const { paymentsSending } = state
     const item = find(paymentsSending, { paymentId })
@@ -456,7 +393,6 @@ const ACTION_HANDLERS = {
       }
     }
   },
-
   [PAYMENT_SUCCESSFUL]: (state, { paymentId }) => {
     const { paymentsSending } = state
     const item = find(paymentsSending, { paymentId })
@@ -473,29 +409,5 @@ const ACTION_HANDLERS = {
     }
   },
 }
-
-// ------------------------------------
-// Selectors
-// ------------------------------------
-
-const paymentSelectors = {}
-const modalPaymentSelector = state => state.payment.payment
-const paymentsSelector = state => state.payment.payments
-const paymentsSendingSelector = state => state.payment.paymentsSending
-const nodesSelector = state => networkSelectors.nodes(state)
-
-paymentSelectors.payments = createSelector(paymentsSelector, nodesSelector, (payments, nodes) =>
-  payments.map(payment => decoratePayment(payment, nodes))
-)
-
-paymentSelectors.paymentsSending = createSelector(
-  paymentsSendingSelector,
-  nodesSelector,
-  (paymentsSending, nodes) => paymentsSending.map(payment => decoratePayment(payment, nodes))
-)
-
-paymentSelectors.paymentModalOpen = createSelector(modalPaymentSelector, payment => !!payment)
-
-export { paymentSelectors }
 
 export default createReducer(initialState, ACTION_HANDLERS)
